Fixes adding subfolders as individual packages
It is entirely possible that prior improper builds that are cached
get used in the build process. The current configuration as
provided in the pyproject.toml file is now correct at ensuring
subfolders do not become their own individual packages. This has
been tested against a clean venv.
In the event of any issues, please clean out any temporary files.